feat: Export Slippi 3.14 data to json
In https://github.com/pcrain/slippc/commit/bf6fd6089cfce0c37ea583fca3c756b793a87758, the ability to parse out the new data like `match_id` was added. However, this data was never added to the json outputs anywhere, so if you ran `slippc -j` there was no way to read this data from external sources.

This PR adds this data to the JSON output so that it can be consumed.
For example, the new `match_id` is going to be very handy for splitting
between ranked and unranked replays.