ci: Run tests and build on every PR and push
I've written a minimal Dockerfile that can successfully run the tests
and build the slippc binary. By having the build itself done inside of a
Dockerfile, the build environment is portable to being runnable by
anyone who can run a docker container build.

At the same time, I've created a minimal GitHub Action workflow that
exercises said docker build so that every time there's a pull request or
a merge to the default branch, the tests will be run.